+++ activerecord/CHANGELOG.md
@@ -1,3 +1,7 @@
+*   The PostgreSQL adapter supports custom domains. Fixes #14305.
+
+    *Yves Senn*
+
 *   PostgreSQL `Column#type` is now determined through the corresponding OID.
     The column types stay the same except for enum columns. They no longer have
     `nil` as type but `enum`.